pyproject.toml
[project]
name = "wethervane"
version = "0.1.0"
description = "Community-covariance model of U.S. political behavior"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # Data handling
    "pandas>=2.0",
    "geopandas>=0.14",
    "pyarrow>=14.0",
    # Census/ACS access
    "cenpy>=1.0",
    # Community detection
    "leidenalg>=0.10",
    "python-igraph>=0.11",
    "infomap>=2.7",
    # Spatial community discovery
    "libpysal>=4.9",
    "networkx>=3.0",
    "kneed>=0.8",
    # Machine learning / matrix factorization
    "scikit-learn>=1.4",
    # lightfm removed: fails to build on Python 3.13 (old C extension setup.py)
    # Re-evaluate at Stage 2; NMF via scikit-learn is the primary approach
    # Bayesian modeling
    "cmdstanpy>=1.2",
    "pymc>=5.10",
    "arviz>=0.17",
    # Graph signal processing
    "pygsp>=0.5",
    # Visualization
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "plotly>=5.18",
    # Utilities
    "requests>=2.31",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
    "openpyxl>=3.1.5",
    "lxml>=5.0",
    "beautifulsoup4>=4.12",
    "pyreadr>=0.5.4",
    "pyyaml>=6.0",
    "duckdb==1.5.0",
    "pydantic>=2.0",
    "hdbscan>=0.8.41",
    "httpx>=0.28.1",
    "ruff>=0.15.6",
    "pdfplumber>=0.11.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "ruff>=0.3",
    "jupyter>=1.0",
    "ipykernel>=6.0",
    "streamlit>=1.30",
]

api = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
]

graph-tool = [
    # graph-tool must be installed via conda or system package manager
    # It cannot be installed via pip
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests", "api/tests"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "fastapi>=0.110.0",
    "httpx>=0.27.0",
    "pytest>=9.0.2",
    "uvicorn[standard]>=0.29.0",
]
